+++ tools-boot-starter/src/main/java/top/zopx/starter/tools/basic/Pagination.java
@@ -19,37 +19,37 @@ public class Pagination implements Serializable {
     private static final long serialVersionUID = 1L;
 
     //当前显示的条数,默认显示20条
-    private int pageSize = 20;
+    private long pageSize = 20L;
 
     //当前显示的页数
-    private int currentIndex;
+    private long currentIndex;
 
     //查询结果总条数
-    private int totalCount;
+    private long totalCount;
 
     public Pagination(int pageSize, int currentIndex) {
         this.pageSize = pageSize;
         this.currentIndex = currentIndex;
     }
 
-    int getStartPage() {
+    long getStartPage() {
         if (this.currentIndex <= 0)
             this.currentIndex = 1;
         if (this.currentIndex >= getTotalPage())
             this.currentIndex = getTotalPage();
-        int startPage = (this.currentIndex - 1) * this.pageSize;
+        long startPage = (this.currentIndex - 1) * this.pageSize;
         return startPage < 0 ? 0 : startPage;
     }
 
-    int getTotalPage() {
+    long getTotalPage() {
         return this.totalCount % this.pageSize != 0 ?    //
                 this.totalCount / this.pageSize + 1 :            //
                 this.totalCount / this.pageSize;
     }
 
-    public int getPageSize() {
-        if (pageSize > 200)
-            return 200;
+    public long getPageSize() {
+        if (pageSize > 200L)
+            return 200L;
         return pageSize;
     }
 }
